[stable23] Allow to disable AuthToken v1

+++ lib/private/Authentication/Token/DefaultTokenProvider.php
@@ -80,6 +80,10 @@ class DefaultTokenProvider implements IProvider {
 								  string $name,
 								  int $type = IToken::TEMPORARY_TOKEN,
 								  int $remember = IToken::DO_NOT_REMEMBER): IToken {
+		if ($this->config->getSystemValueBool('auth.authtoken.v1.disabled')) {
+			throw new InvalidTokenException('Authtokens v1 disabled');
+		}
+
 		$dbToken = new DefaultToken();
 		$dbToken->setUid($uid);
 		$dbToken->setLoginName($loginName);
@@ -106,6 +110,10 @@ class DefaultTokenProvider implements IProvider {
 	 * @throws InvalidTokenException
 	 */
 	public function updateToken(IToken $token) {
+		if ($this->config->getSystemValueBool('auth.authtoken.v1.disabled')) {
+			throw new InvalidTokenException('Authtokens v1 disabled');
+		}
+
 		if (!($token instanceof DefaultToken)) {
 			throw new InvalidTokenException("Invalid token type");
 		}
@@ -119,6 +127,10 @@ class DefaultTokenProvider implements IProvider {
 	 * @param IToken $token
 	 */
 	public function updateTokenActivity(IToken $token) {
+		if ($this->config->getSystemValueBool('auth.authtoken.v1.disabled')) {
+			throw new InvalidTokenException('Authtokens v1 disabled');
+		}
+
 		if (!($token instanceof DefaultToken)) {
 			throw new InvalidTokenException("Invalid token type");
 		}
@@ -132,6 +144,10 @@ class DefaultTokenProvider implements IProvider {
 	}
 
 	public function getTokenByUser(string $uid): array {
+		if ($this->config->getSystemValueBool('auth.authtoken.v1.disabled')) {
+			return [];
+		}
+
 		return $this->mapper->getTokenByUser($uid);
 	}
 
@@ -144,6 +160,10 @@ class DefaultTokenProvider implements IProvider {
 	 * @return IToken
 	 */
 	public function getToken(string $tokenId): IToken {
+		if ($this->config->getSystemValueBool('auth.authtoken.v1.disabled')) {
+			throw new InvalidTokenException('Authtokens v1 disabled');
+		}
+
 		try {
 			$token = $this->mapper->getToken($this->hashToken($tokenId));
 		} catch (DoesNotExistException $ex) {
@@ -166,6 +186,10 @@ class DefaultTokenProvider implements IProvider {
 	 * @return IToken
 	 */
 	public function getTokenById(int $tokenId): IToken {
+		if ($this->config->getSystemValueBool('auth.authtoken.v1.disabled')) {
+			throw new InvalidTokenException('Authtokens v1 disabled');
+		}
+
 		try {
 			$token = $this->mapper->getTokenById($tokenId);
 		} catch (DoesNotExistException $ex) {
@@ -186,6 +210,10 @@ class DefaultTokenProvider implements IProvider {
 	 * @return IToken
 	 */
 	public function renewSessionToken(string $oldSessionId, string $sessionId): IToken {
+		if ($this->config->getSystemValueBool('auth.authtoken.v1.disabled')) {
+			throw new InvalidTokenException('Authtokens v1 disabled');
+		}
+
 		$token = $this->getToken($oldSessionId);
 
 		$newToken = new DefaultToken();
@@ -214,6 +242,10 @@ class DefaultTokenProvider implements IProvider {
 	 * @return string
 	 */
 	public function getPassword(IToken $savedToken, string $tokenId): string {
+		if ($this->config->getSystemValueBool('auth.authtoken.v1.disabled')) {
+			throw new InvalidTokenException('Authtokens v1 disabled');
+		}
+
 		$password = $savedToken->getPassword();
 		if ($password === null || $password === '') {
 			throw new PasswordlessTokenException();
@@ -230,6 +262,10 @@ class DefaultTokenProvider implements IProvider {
 	 * @throws InvalidTokenException
 	 */
 	public function setPassword(IToken $token, string $tokenId, string $password) {
+		if ($this->config->getSystemValueBool('auth.authtoken.v1.disabled')) {
+			throw new InvalidTokenException('Authtokens v1 disabled');
+		}
+
 		if (!($token instanceof DefaultToken)) {
 			throw new InvalidTokenException("Invalid token type");
 		}
@@ -244,10 +280,18 @@ class DefaultTokenProvider implements IProvider {
 	 * @param string $token
 	 */
 	public function invalidateToken(string $token) {
+		if ($this->config->getSystemValueBool('auth.authtoken.v1.disabled')) {
+			return;
+		}
+
 		$this->mapper->invalidate($this->hashToken($token));
 	}
 
 	public function invalidateTokenById(string $uid, int $id) {
+		if ($this->config->getSystemValueBool('auth.authtoken.v1.disabled')) {
+			return;
+		}
+
 		$this->mapper->deleteById($uid, $id);
 	}
 
@@ -255,6 +299,10 @@ class DefaultTokenProvider implements IProvider {
 	 * Invalidate (delete) old session tokens
 	 */
 	public function invalidateOldTokens() {
+		if ($this->config->getSystemValueBool('auth.authtoken.v1.disabled')) {
+			return;
+		}
+
 		$olderThan = $this->time->getTime() - (int) $this->config->getSystemValue('session_lifetime', 60 * 60 * 24);
 		$this->logger->debug('Invalidating session tokens older than ' . date('c', $olderThan), ['app' => 'cron']);
 		$this->mapper->invalidateOld($olderThan, IToken::DO_NOT_REMEMBER);
@@ -272,6 +320,10 @@ class DefaultTokenProvider implements IProvider {
 	 * @return IToken
 	 */
 	public function rotate(IToken $token, string $oldTokenId, string $newTokenId): IToken {
+		if ($this->config->getSystemValueBool('auth.authtoken.v1.disabled')) {
+			throw new InvalidTokenException('Authtokens v1 disabled');
+		}
+
 		try {
 			$password = $this->getPassword($token, $oldTokenId);
 			$token->setPassword($this->encryptPassword($password, $newTokenId));
@@ -329,6 +381,10 @@ class DefaultTokenProvider implements IProvider {
 	}
 
 	public function markPasswordInvalid(IToken $token, string $tokenId) {
+		if ($this->config->getSystemValueBool('auth.authtoken.v1.disabled')) {
+			throw new InvalidTokenException('Authtokens v1 disabled');
+		}
+
 		if (!($token instanceof DefaultToken)) {
 			throw new InvalidTokenException("Invalid token type");
 		}
